feat(legends): migrate package to typescript
Close #1221
