--- packages/legends/src/index.js
@@ -1,13 +1,0 @@
-/*
- * This file is part of the nivo project.
- *
- * Copyright 2016-present, Raphaël Benitte.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-export * from './svg'
-export * from './canvas'
-export * from './constants'
-export * from './props'
-export * from './hooks'
