--- packages/legends/src/svg/LegendSvg.js
@@ -1,142 +1,0 @@
-/*
- * This file is part of the nivo project.
- *
- * Copyright 2016-present, Raphaël Benitte.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-import React from 'react'
-import PropTypes from 'prop-types'
-import LegendSvgItem from './LegendSvgItem'
-import { datumPropType, symbolPropTypes, interactivityPropTypes } from '../props'
-import { computeDimensions } from '../compute'
-import {
-    DIRECTION_LEFT_TO_RIGHT,
-    DIRECTION_RIGHT_TO_LEFT,
-    DIRECTION_TOP_TO_BOTTOM,
-    DIRECTION_BOTTOM_TO_TOP,
-} from '../constants'
-
-const LegendSvg = ({
-    data,
-
-    x,
-    y,
-    direction,
-    padding: _padding,
-    justify,
-    effects,
-
-    itemWidth,
-    itemHeight,
-    itemDirection,
-    itemsSpacing,
-    itemTextColor,
-    itemBackground,
-    itemOpacity,
-
-    symbolShape,
-    symbolSize,
-    symbolSpacing,
-    symbolBorderWidth,
-    symbolBorderColor,
-
-    onClick,
-    onMouseEnter,
-    onMouseLeave,
-}) => {
-    // eslint-disable-next-line no-unused-vars
-    const { width, height, padding } = computeDimensions({
-        itemCount: data.length,
-        itemWidth,
-        itemHeight,
-        itemsSpacing,
-        direction,
-        padding: _padding,
-    })
-
-    let xStep = 0
-    let yStep = 0
-    if (direction === 'row') {
-        xStep = itemWidth + itemsSpacing
-    } else if (direction === 'column') {
-        yStep = itemHeight + itemsSpacing
-    }
-
-    return (
-        <g transform={`translate(${x},${y})`}>
-            {data.map((data, i) => (
-                <LegendSvgItem
-                    key={i}
-                    data={data}
-                    x={i * xStep + padding.left}
-                    y={i * yStep + padding.top}
-                    width={itemWidth}
-                    height={itemHeight}
-                    direction={itemDirection}
-                    justify={justify}
-                    effects={effects}
-                    textColor={itemTextColor}
-                    background={itemBackground}
-                    opacity={itemOpacity}
-                    symbolShape={symbolShape}
-                    symbolSize={symbolSize}
-                    symbolSpacing={symbolSpacing}
-                    symbolBorderWidth={symbolBorderWidth}
-                    symbolBorderColor={symbolBorderColor}
-                    onClick={onClick}
-                    onMouseEnter={onMouseEnter}
-                    onMouseLeave={onMouseLeave}
-                />
-            ))}
-        </g>
-    )
-}
-
-LegendSvg.propTypes = {
-    data: PropTypes.arrayOf(datumPropType).isRequired,
-
-    x: PropTypes.number.isRequired,
-    y: PropTypes.number.isRequired,
-    direction: PropTypes.oneOf(['row', 'column']).isRequired,
-    padding: PropTypes.oneOfType([
-        PropTypes.number,
-        PropTypes.shape({
-            top: PropTypes.number,
-            right: PropTypes.number,
-            bottom: PropTypes.number,
-            left: PropTypes.number,
-        }),
-    ]).isRequired,
-    justify: PropTypes.bool.isRequired,
-
-    itemsSpacing: PropTypes.number.isRequired,
-    itemWidth: PropTypes.number.isRequired,
-    itemHeight: PropTypes.number.isRequired,
-    itemDirection: PropTypes.oneOf([
-        DIRECTION_LEFT_TO_RIGHT,
-        DIRECTION_RIGHT_TO_LEFT,
-        DIRECTION_TOP_TO_BOTTOM,
-        DIRECTION_BOTTOM_TO_TOP,
-    ]).isRequired,
-    itemTextColor: PropTypes.string.isRequired,
-    itemBackground: PropTypes.string.isRequired,
-    itemOpacity: PropTypes.number.isRequired,
-
-    ...symbolPropTypes,
-    ...interactivityPropTypes,
-}
-
-LegendSvg.defaultProps = {
-    padding: 0,
-    justify: false,
-
-    itemsSpacing: 0,
-    itemDirection: 'left-to-right',
-    itemTextColor: 'black',
-    itemBackground: 'transparent',
-    itemOpacity: 1,
-}
-
-export default LegendSvg
