+++ packages/legends/src/svg/index.ts
@@ -1,0 +1,3 @@
+export * from './BoxLegendSvg'
+export * from './LegendSvg'
+export * from './LegendSvgItem'
