--- packages/legends/src/svg/symbols/index.js
@@ -1,12 +1,0 @@
-/*
- * This file is part of the nivo project.
- *
- * Copyright 2016-present, Raphaël Benitte.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-export { default as SymbolCircle } from './SymbolCircle'
-export { default as SymbolDiamond } from './SymbolDiamond'
-export { default as SymbolSquare } from './SymbolSquare'
-export { default as SymbolTriangle } from './SymbolTriangle'
