+++ packages/legends/src/svg/symbols/index.ts
@@ -1,0 +1,4 @@
+export * from './SymbolCircle'
+export * from './SymbolDiamond'
+export * from './SymbolSquare'
+export * from './SymbolTriangle'
