+++ packages/legends/src/svg/symbols/types.ts
@@ -1,0 +1,9 @@
+export type SymbolProps = {
+    id: string | number
+    x: number
+    y: number
+    size: number
+    fill: string
+    borderWidth?: number
+    borderColor?: string
+}
